karm-gfx: Fixed some strokes glitches.

+++ src/libs/karm-gfx/context.h
@@ -203,7 +203,7 @@ struct Context {
 
     /* --- Effects ---------------------------------------------------------- */
 
-    void blur(Math::Recti region, double radius);
+    void blur(Math::Recti region, int radius);
 
     void saturate(Math::Recti region, double saturation);
